+++ radsat-sk/operation/subsystems/RCamera.h
@@ -23,12 +23,12 @@
 
 /* Struct used to define ADCS Function*/
 typedef struct _detection_results_t {
-	uint16_t sunSensorX;
-	uint16_t sunSensorY;
-	uint16_t sunSensorZ;
-	uint16_t imageSensorX;
-	uint16_t imageSensorY;
-	uint16_t imageSensorZ;
+	float sunSensorX;
+	float sunSensorY;
+	float sunSensorZ;
+	float imageSensorX;
+	float imageSensorY;
+	float imageSensorZ;
 } detection_results_t;
 
 /* Struct that holds all power related telemetry */
